[Chore] rename to `src/workbench/services`
