+++ src/code/browser/renderer.ts
@@ -2,13 +2,13 @@ import "src/styles/index.scss";
 import "src/code/common/common.register";
 import "src/workbench/parts/workspace/editor/editor";
 import { Workbench } from "src/workbench/workbench";
-import { workbenchShortcutRegistrations } from "src/workbench/service/workbench/shortcut.register";
-import { workbenchCommandRegistrations } from "src/workbench/service/workbench/command.register";
+import { workbenchShortcutRegistrations } from "src/workbench/services/workbench/shortcut.register";
+import { workbenchCommandRegistrations } from "src/workbench/services/workbench/command.register";
 import { rendererServiceRegistrations } from "src/code/browser/service.register";
 import { IInstantiationService, InstantiationService } from "src/platform/instantiation/common/instantiation";
 import { getSingletonServiceDescriptors, ServiceCollection } from "src/platform/instantiation/common/serviceCollection";
 import { waitDomToBeLoad } from "src/base/browser/basic/dom";
-import { ComponentService, IComponentService } from "src/workbench/service/component/componentService";
+import { ComponentService, IComponentService } from "src/workbench/services/component/componentService";
 import { Disposable } from "src/base/common/dispose";
 import { ServiceDescriptor } from "src/platform/instantiation/common/descriptor";
 import { initExposedElectronAPIs } from "src/platform/electron/browser/global";
