+++ src/code/browser/service.register.ts
@@ -1,19 +1,19 @@
 import { ServiceDescriptor } from "src/platform/instantiation/common/descriptor";
 import { registerSingleton } from "src/platform/instantiation/common/serviceCollection";
 import { CommandService, ICommandService } from "src/platform/command/common/commandService";
-import { IKeyboardScreenCastService, KeyboardScreenCastService } from "src/workbench/service/keyboard/keyboardScreenCastService";
-import { IKeyboardService, KeyboardService } from "src/workbench/service/keyboard/keyboardService";
+import { IKeyboardScreenCastService, KeyboardScreenCastService } from "src/workbench/services/keyboard/keyboardScreenCastService";
+import { IKeyboardService, KeyboardService } from "src/workbench/services/keyboard/keyboardService";
 import { BrowserDialogService, IDialogService } from 'src/platform/dialog/browser/browserDialogService';
-import { IShortcutService, ShortcutService } from 'src/workbench/service/shortcut/shortcutService';
+import { IShortcutService, ShortcutService } from 'src/workbench/services/shortcut/shortcutService';
 import { SideBar, ISideBarService } from 'src/workbench/parts/sideBar/sideBar';
 import { IWorkspaceService, WorkspaceComponent } from 'src/workbench/parts/workspace/workspace';
 import { SideViewService, ISideViewService } from 'src/workbench/parts/sideView/sideView';
-import { ExplorerTreeService, IExplorerTreeService } from 'src/workbench/service/explorerTree/explorerTreeService';
+import { ExplorerTreeService, IExplorerTreeService } from 'src/workbench/services/explorerTree/explorerTreeService';
 import { ContextService, IContextService } from 'src/platform/context/common/contextService';
-import { ContextMenuService, IContextMenuService } from "src/workbench/service/contextMenu/contextMenuService";
-import { ILayoutService, LayoutService } from "src/workbench/service/layout/layoutService";
-import { INotificationService, NotificationService } from "src/workbench/service/notification/notificationService";
-import { IThemeService, ThemeService } from "src/workbench/service/theme/themeService";
+import { ContextMenuService, IContextMenuService } from "src/workbench/services/contextMenu/contextMenuService";
+import { ILayoutService, LayoutService } from "src/workbench/services/layout/layoutService";
+import { INotificationService, NotificationService } from "src/workbench/services/notification/notificationService";
+import { IThemeService, ThemeService } from "src/workbench/services/theme/themeService";
 
 /*******************************************************************************
  * Registraion for desktop browser-side non-important microservices.
