+++ src/workbench/services/classicTree/classicRenderer.ts
@@ -2,7 +2,7 @@ import { ITreeNode } from "src/base/browser/secondary/tree/tree";
 import { ITreeListRenderer } from "src/base/browser/secondary/tree/treeListRenderer";
 import { IListViewMetadata, RendererType } from "src/base/browser/secondary/listView/listRenderer";
 import { FuzzyScore } from "src/base/common/fuzzy";
-import { ClassicItem } from "src/workbench/service/classicTree/classicItem";
+import { ClassicItem } from "src/workbench/services/classicTree/classicItem";
 import { Icons } from "src/base/browser/icon/icons";
 import { getIconClass } from "src/base/browser/icon/iconRegistry";
 import { IListItemProvider } from "src/base/browser/secondary/listView/listItemProvider";
