+++ src/workbench/services/notebookTree/notebookTreeService.ts
@@ -1,7 +1,7 @@
 import { Disposable } from "src/base/common/dispose";
 import { Register } from "src/base/common/event";
 import { URI } from "src/base/common/file/uri";
-import { ITreeService, TreeMode } from "src/workbench/service/explorerTree/treeService";
+import { ITreeService, TreeMode } from "src/workbench/services/explorerTree/treeService";
 
 export interface INotebookTreeService extends ITreeService<any> {
