+++ src/workbench/services/shortcut/shortcutRegistrant.ts
