+++ src/workbench/services/shortcut/shortcutService.ts
@@ -2,12 +2,12 @@ import { Disposable, IDisposable } from "src/base/common/dispose";
 import { DataBuffer } from "src/base/common/file/buffer";
 import { URI } from "src/base/common/file/uri";
 import { Shortcut, ShortcutHash } from "src/base/common/keyboard";
-import { IKeyboardService } from "src/workbench/service/keyboard/keyboardService";
+import { IKeyboardService } from "src/workbench/services/keyboard/keyboardService";
 import { IFileService } from "src/platform/files/common/fileService";
 import { IService, createService } from "src/platform/instantiation/common/decorator";
 import { ILogService } from "src/base/common/logger";
 import { IBrowserLifecycleService, ILifecycleService, LifecyclePhase } from "src/platform/lifecycle/browser/browserLifecycleService";
-import { IShortcutItem, IShortcutRegistrant, IShortcutRegistration, IShortcutWithCommandRegistration, ShortcutWeight } from "src/workbench/service/shortcut/shortcutRegistrant";
+import { IShortcutItem, IShortcutRegistrant, IShortcutRegistration, IShortcutWithCommandRegistration, ShortcutWeight } from "src/workbench/services/shortcut/shortcutRegistrant";
 import { REGISTRANTS } from "src/platform/registrant/common/registrant";
 import { IBrowserEnvironmentService } from "src/platform/environment/common/environment";
 import { Emitter, Register } from "src/base/common/event";
