+++ src/workbench/services/workbench/command.register.ts
