+++ src/workbench/services/workbench/workbenchService.ts
@@ -1,4 +1,4 @@
-import { IComponent } from "src/workbench/service/component/component";
+import { IComponent } from "src/workbench/services/component/component";
 import { IService, createService } from "src/platform/instantiation/common/decorator";
 
 export const IWorkbenchService = createService<IWorkbenchService>('workbench-service');
