+++ src/workbench/workbench.ts
@@ -1,9 +1,9 @@
 import 'src/workbench/media/workbench.scss';
 import { IInstantiationService } from "src/platform/instantiation/common/instantiation";
-import { IComponentService } from "src/workbench/service/component/componentService";
+import { IComponentService } from "src/workbench/services/component/componentService";
 import { WorkbenchLayout } from "src/workbench/layout";
-import { IWorkbenchService } from "src/workbench/service/workbench/workbenchService";
-import { IKeyboardScreenCastService } from "src/workbench/service/keyboard/keyboardScreenCastService";
+import { IWorkbenchService } from "src/workbench/services/workbench/workbenchService";
+import { IKeyboardScreenCastService } from "src/workbench/services/keyboard/keyboardScreenCastService";
 import { ISideBarService } from "src/workbench/parts/sideBar/sideBar";
 import { ISideViewService } from "src/workbench/parts/sideView/sideView";
 import { IWorkspaceService } from "src/workbench/parts/workspace/workspace";
@@ -13,9 +13,9 @@ import { IContextKey } from 'src/platform/context/common/contextKey';
 import { IS_LINUX, IS_MAC, IS_WINDOWS } from 'src/base/common/platform';
 import { IBrowserLifecycleService, ILifecycleService, LifecyclePhase } from 'src/platform/lifecycle/browser/browserLifecycleService';
 import { IBrowserEnvironmentService, IEnvironmentService } from 'src/platform/environment/common/environment';
-import { IContextMenuService } from 'src/workbench/service/contextMenu/contextMenuService';
-import { ILayoutService } from 'src/workbench/service/layout/layoutService';
-import { IThemeService } from 'src/workbench/service/theme/themeService';
+import { IContextMenuService } from 'src/workbench/services/contextMenu/contextMenuService';
+import { ILayoutService } from 'src/workbench/services/layout/layoutService';
+import { IThemeService } from 'src/workbench/services/theme/themeService';
 import { IConfigurationService } from 'src/platform/configuration/common/configuration';
 import { WorkbenchConfiguration } from 'src/code/browser/configuration.register';
 import { SideViewConfiguration } from 'src/workbench/parts/sideView/configuration.register';
